Refactor ArbitraryBuilder as interface
