+++ fixture-monkey-autoparams/src/main/java/com/navercorp/fixturemonkey/autoparams/customization/FixtureMonkeyArbitraryBuilderCustomizer.java
@@ -25,8 +25,8 @@ import org.javaunit.autoparams.customization.Customizer;
 import org.javaunit.autoparams.generator.ObjectContainer;
 import org.javaunit.autoparams.generator.ObjectGenerator;
 
-import com.navercorp.fixturemonkey.ArbitraryBuilder;
 import com.navercorp.fixturemonkey.FixtureMonkey;
+import com.navercorp.fixturemonkey.OldArbitraryBuilderImpl;
 
 final class FixtureMonkeyArbitraryBuilderCustomizer implements Customizer {
 	private final FixtureMonkey fixtureMonkey;
@@ -49,7 +49,7 @@ final class FixtureMonkeyArbitraryBuilderCustomizer implements Customizer {
 
 	private ObjectContainer generate(ParameterizedType parameterizedType) {
 		Class<?> type = (Class<?>)parameterizedType.getRawType();
-		return type == ArbitraryBuilder.class
+		return type == OldArbitraryBuilderImpl.class
 			? this.generate((Class<?>)parameterizedType.getActualTypeArguments()[0])
 			: ObjectContainer.EMPTY;
 	}
