+++ fixture-monkey-jackson/src/test/java/com/navercorp/fixturemonkey/test/FixtureMonkeyJacksonArbitraryGeneratorTest.java
@@ -85,80 +85,6 @@ class FixtureMonkeyJacksonArbitraryGeneratorTest {
 	}
 
 	@Property
-	void giveMeJsonPropertySetGenerator() {
-		// when
-		JsonPropertyClass actual = sut.giveMeBuilder(JsonPropertyClass.class)
-			.generator(BeanArbitraryGenerator.INSTANCE)
-			.set("jsonValue", "set")
-			.sample();
-
-		then(actual.value).isNotEqualTo("set");
-	}
-
-	@Property
-	void giveMeJsonPropertySetGeneratorToJackson() {
-		// when
-		JsonPropertyClass actual = sut.giveMeBuilder(JsonPropertyClass.class)
-			.generator(BeanArbitraryGenerator.INSTANCE)
-			.generator(JacksonArbitraryGenerator.INSTANCE)
-			.set("jsonValue", "set")
-			.sample();
-
-		then(actual.value).isEqualTo("set");
-	}
-
-	@Property
-	void giveMeJsonPropertySetGeneratorToJacksonWithExpressionGenerator() {
-		// given
-		TypeReference<JsonPropertyClass> typeReference = new TypeReference<JsonPropertyClass>() {
-		};
-		ExpressionGenerator expressionGenerator = resolver -> {
-			com.navercorp.fixturemonkey.api.property.Property property =
-				PropertyCache.getProperty(typeReference.getAnnotatedType(), "value").get();
-			return resolver.resolve(property);
-		};
-
-		// when
-		JsonPropertyClass actual = sut.giveMeBuilder(JsonPropertyClass.class)
-			.generator(BeanArbitraryGenerator.INSTANCE)
-			.generator(JacksonArbitraryGenerator.INSTANCE)
-			.set(expressionGenerator, "set")
-			.sample();
-
-		then(actual.value).isEqualTo("set");
-	}
-
-	@Property
-	void giveMeJsonPropertySetFailedAfterDecompose() {
-		// given
-		JsonPropertyClass jsonPropertyClass = new JsonPropertyClass();
-		jsonPropertyClass.setValue("jsonValue");
-
-		// when
-		JsonPropertyClass actual = sut.giveMeBuilder(jsonPropertyClass)
-			.generator(BeanArbitraryGenerator.INSTANCE)
-			.set("jsonValue", "set")
-			.sample();
-
-		then(actual.value).isEqualTo("jsonValue");
-	}
-
-	@Property
-	void giveMeJsonPropertySetAfterDecompose() {
-		// given
-		JsonPropertyClass jsonPropertyClass = new JsonPropertyClass();
-		jsonPropertyClass.setValue("jsonValue");
-
-		// when
-		JsonPropertyClass actual = sut.giveMeBuilder(jsonPropertyClass)
-			.generator(BeanArbitraryGenerator.INSTANCE)
-			.set("value", "set")
-			.sample();
-
-		then(actual.value).isEqualTo("set");
-	}
-
-	@Property
 	void giveMeJsonNodeReturnsNull() {
 		// when
 		JsonNodeWrapperClass actual = sut.giveMeOne(JsonNodeWrapperClass.class);
