+++ fixture-monkey/src/main/java/com/navercorp/fixturemonkey/arbitrary/ArbitrarySetLazyValue.java
@@ -27,7 +27,7 @@ import org.apiguardian.api.API.Status;
 
 import net.jqwik.api.Arbitrary;
 
-import com.navercorp.fixturemonkey.ArbitraryBuilder;
+import com.navercorp.fixturemonkey.OldArbitraryBuilderImpl;
 
 @API(since = "0.4.0", status = Status.EXPERIMENTAL)
 public final class ArbitrarySetLazyValue<T> extends AbstractArbitrarySet<T> {
@@ -46,7 +46,7 @@ public final class ArbitrarySetLazyValue<T> extends AbstractArbitrarySet<T> {
 
 	@SuppressWarnings({"rawtypes"})
 	@Override
-	public void accept(ArbitraryBuilder arbitraryBuilder) {
+	public void accept(OldArbitraryBuilderImpl arbitraryBuilder) {
 		ArbitraryExpression arbitraryExpression = this.getArbitraryExpression();
 		T value = this.getApplicableValue();
 		long limit = this.limit;
