+++ fixture-monkey/src/test/java/com/navercorp/fixturemonkey/test/FixtureMonkeyTest.java
@@ -1010,250 +1010,6 @@ class FixtureMonkeyTest {
 	}
 
 	@Property
-	void isDirtyReturnsFalse() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	void isDirtyReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenManipulatedAndSampledReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test");
-		arbitraryBuilder.sample();
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenManipulatedAndFixedReturnsFalse() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test")
-			.fixed();
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	void isDirtyWhenManipulatedAndFixedAndManipulatedReturnsFalse() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test")
-			.fixed()
-			.set("value", "fixed");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenManipulatedAndApplyAndManipulatedReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test")
-			.apply((builder, it) -> {
-			})
-			.set("value", "test");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenManipulatedAndAcceptIfAndManipulatedReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test")
-			.acceptIf(it -> true, it -> {
-			})
-			.set("value", "test");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenApplyAndFixedAndManipulatedReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.apply((builder, it) -> {
-			})
-			.fixed()
-			.set("value", "test");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenApplyAndFixedReturnsFalse() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.apply((builder, it) -> {
-			})
-			.fixed();
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	void isDirtyWhenAcceptIfAndFixedReturnsFalse() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.acceptIf(it -> true, it -> {
-			})
-			.fixed();
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	@Domain(FixtureMonkeyTestSpecs.class)
-	void isDirtyWhenDecomposeReturnsFalse(@ForAll StringWithNotBlank stringWithNotBlank) {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(stringWithNotBlank);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	void isDirtyWhenDecomposeAndManipulatedReturnsTrue() {
-		// given
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = SUT.giveMeBuilder(StringWithNotBlank.class)
-			.set("value", "test");
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenRegisterReturnsFalse() {
-		// given
-		FixtureMonkey sut = FixtureMonkey.builder()
-			.register(StringWithNotBlank.class, fixture -> fixture.giveMeBuilder(StringWithNotBlank.class))
-			.build();
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = sut.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	void isDirtyWhenRegisterWithManipulatedReturnsTrue() {
-		// given
-		FixtureMonkey sut = FixtureMonkey.builder()
-			.register(StringWithNotBlank.class, fixture ->
-				fixture.giveMeBuilder(StringWithNotBlank.class)
-					.set("value", "test")
-			)
-			.build();
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = sut.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
-	void isDirtyWhenRegisterWithManipulatedAndFixedReturnsFalse() {
-		// given
-		FixtureMonkey sut = FixtureMonkey.builder()
-			.register(StringWithNotBlank.class, fixture ->
-				fixture.giveMeBuilder(StringWithNotBlank.class)
-					.set("value", "test")
-					.fixed()
-			)
-			.build();
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = sut.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	@Domain(FixtureMonkeyTestSpecs.class)
-	void isDirtyWhenRegisterWithDecomposedReturnsFalse(@ForAll StringWithNotBlank stringWithNotBlank) {
-		// given
-		FixtureMonkey sut = FixtureMonkey.builder()
-			.register(StringWithNotBlank.class, fixture ->
-				fixture.giveMeBuilder(stringWithNotBlank)
-			)
-			.build();
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = sut.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isFalse();
-	}
-
-	@Property
-	@Domain(FixtureMonkeyTestSpecs.class)
-	void isDirtyWhenRegisterWithDecomposedAndManipulatedReturnsTrue(@ForAll StringWithNotBlank stringWithNotBlank) {
-		// given
-		FixtureMonkey sut = FixtureMonkey.builder()
-			.register(StringWithNotBlank.class, fixture ->
-				fixture.giveMeBuilder(stringWithNotBlank)
-					.set("value", "test")
-			)
-			.build();
-		ArbitraryBuilder<StringWithNotBlank> arbitraryBuilder = sut.giveMeBuilder(StringWithNotBlank.class);
-
-		// when
-		boolean changed = arbitraryBuilder.isDirty();
-
-		then(changed).isTrue();
-	}
-
-	@Property
 	@Domain(FixtureMonkeyTestSpecs.class)
 	void giveMeListWithAnnotation(@ForAll ListWithAnnotation actual) {
 		then(actual.getValues()).isNotEmpty();
